+++ backend/canisters/notifications_index/impl/src/jobs/upgrade_canisters.rs
@@ -2,7 +2,7 @@ use crate::{mutate_state, RuntimeState};
 use ic_cdk::timer::TimerId;
 use std::cell::Cell;
 use std::time::Duration;
-use tracing::info;
+use tracing::trace;
 use types::{CanisterId, Version};
 use utils::canister::{upgrade, FailedUpgrade};
 
@@ -19,7 +19,7 @@ pub(crate) fn start_job_if_required(runtime_state: &RuntimeState) -> bool {
     {
         let timer_id = ic_cdk::timer::set_timer_interval(Duration::from_secs(2), run);
         TIMER_ID.with(|t| t.set(Some(timer_id)));
-        info!("'upgrade_canisters' job started");
+        trace!("'upgrade_canisters' job started");
         true
     } else {
         false
@@ -33,7 +33,7 @@ fn run() {
         GetNextResult::QueueEmpty => {
             if let Some(timer_id) = TIMER_ID.with(|t| t.take()) {
                 ic_cdk::timer::clear_timer(timer_id);
-                info!("'upgrade_canisters' job stopped");
+                trace!("'upgrade_canisters' job stopped");
             }
         }
     }