+++ backend/canisters/group_index/CHANGELOG.md
@@ -5,3 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+
+### Changed
+
+- Reduce log level of job started / stopped messages ([#2951](https://github.com/open-ic/open-chat/pull/2951))